refactor: extract shared _refresh_node, fix agent inbox review issues
- Extract _refresh_node into lib/_org_utils.py, import from both
  message_store.py and agent_inbox.py (Critical 1)
- Fix AgentInbox._reload to call _refresh_live_nodes() after reload,
  add generation contract docstring (Critical 2)
- Add "archived" key to AgentInbox.counts() (Important 3)
- Add test_retry_increments_count_twice asserting RETRY_COUNT == 2
  (Important 5)
